refactor: move codes to store
