+++ apps/web/src/services/dashboards/components/dashboard-folder/DashboardFolderCloneModal.vue
@@ -1,26 +1,13 @@
 <script lang="ts" setup>
 import { computed, reactive, watch } from 'vue';
 
-import { SpaceConnector } from '@cloudforet/core-lib/space-connector';
 import {
     PDataTable, PI, PToggleButton, PButtonModal,
 } from '@cloudforet/mirinae';
-import { getClonedName } from '@cloudforet/utils';
 
-import type { ListResponse } from '@/schema/_common/api-verbs/list';
-import { RESOURCE_GROUP } from '@/schema/_common/constant';
-import type { DashboardCreateParams } from '@/schema/dashboard/_types/dashboard-type';
-import type { FolderCreateParams } from '@/schema/dashboard/_types/folder-type';
-import type { WidgetModel, WidgetListParams } from '@/schema/dashboard/_types/widget-type';
-import type { PublicDashboardCreateParameters } from '@/schema/dashboard/public-dashboard/api-verbs/create';
-import type { PublicFolderCreateParameters } from '@/schema/dashboard/public-folder/api-verbs/create';
-import { store } from '@/store';
 import { i18n } from '@/translations';
 
-import { useAppContextStore } from '@/store/app-context/app-context-store';
 import { useDashboardStore } from '@/store/dashboard/dashboard-store';
-import { useAllReferenceStore } from '@/store/reference/all-reference-store';
-import type { CostDataSourceReferenceMap } from '@/store/reference/cost-data-source-reference-store';
 
 import { showSuccessMessage } from '@/lib/helper/notice-alert-helper';
 
@@ -29,9 +16,8 @@ import { useProxyValue } from '@/common/composables/proxy-state';
 
 import { gray } from '@/styles/colors';
 
-import { getSharedDashboardLayouts } from '@/services/dashboards/helpers/dashboard-share-helper';
 import { useDashboardMainPageStore } from '@/services/dashboards/stores/dashboard-main-page-store';
-import type { DashboardTreeDataType, DashboardDataTableItem } from '@/services/dashboards/types/dashboard-folder-type';
+import type { DashboardDataTableItem } from '@/services/dashboards/types/dashboard-folder-type';
 
 
 
@@ -48,16 +34,10 @@ const props = withDefaults(defineProps<Props>(), {
 });
 const emit = defineEmits<{(e: 'update:visible', visible: boolean): void,
 }>();
-const appContextStore = useAppContextStore();
-const allReferenceStore = useAllReferenceStore();
 const dashboardStore = useDashboardStore();
 const dashboardMainPageStore = useDashboardMainPageStore();
 const dashboardMainPageState = dashboardMainPageStore.state;
 const dashboardMainPageGetters = dashboardMainPageStore.getters;
-const storeState = reactive({
-    isAdminMode: computed(() => appContextStore.getters.isAdminMode),
-    costDataSource: computed<CostDataSourceReferenceMap>(() => allReferenceStore.getters.costDataSource),
-});
 const state = reactive({
     loading: false,
     proxyVisible: useProxyValue<boolean>('visible', props, emit),
@@ -77,67 +57,14 @@ const state = reactive({
 });
 
 /* Api */
-const listDashboardWidgets = async (dashboardId: string): Promise<WidgetModel[]> => {
-    try {
-        const isPrivate = dashboardId.startsWith('private');
-        const fetcher = isPrivate
-            ? SpaceConnector.clientV2.dashboard.privateWidget.list
-            : SpaceConnector.clientV2.dashboard.publicWidget.list;
-        const { results } = await fetcher<WidgetListParams, ListResponse<WidgetModel>>({
-            dashboard_id: dashboardId,
-        });
-        return results || [];
-    } catch (e) {
-        ErrorHandler.handleError(e);
-        return [];
-    }
-};
-const createFolder = async (originFolderName: string, isPrivate: boolean): Promise<string|undefined> => {
-    try {
-        const fetcher = isPrivate ? SpaceConnector.clientV2.dashboard.privateFolder.create : SpaceConnector.clientV2.dashboard.publicFolder.create;
-        const params: FolderCreateParams = {
-            name: getClonedName(dashboardMainPageGetters.existingFolderNameList, originFolderName),
-            tags: { created_by: store.state.user.userId },
-        };
-        if (!isPrivate) {
-            (params as PublicFolderCreateParameters).resource_group = storeState.isAdminMode ? RESOURCE_GROUP.DOMAIN : RESOURCE_GROUP.WORKSPACE;
-        }
-        const createdFolder = await fetcher(params);
+const cloneDashboard = async (dashboardId: string, isPrivate?: boolean, folderId?: string) => {
+    const createdDashboard = await dashboardStore.cloneDashboard(dashboardId, isPrivate, folderId);
+    if (createdDashboard) {
         dashboardMainPageStore.setNewIdList([
             ...dashboardMainPageState.newIdList,
-            createdFolder.folder_id as string,
+            createdDashboard.dashboard_id as string,
         ]);
-        return createdFolder.folder_id;
-    } catch (e) {
-        return undefined;
     }
-};
-const createDashboard = async (treeData: DashboardTreeDataType, isPrivate?: boolean, folderId?: string) => {
-    const _dashboardType = isPrivate ? 'PRIVATE' : 'PUBLIC';
-    const _isOriginDashboardPrivate = treeData.id.startsWith('private');
-    const _dashboard = _isOriginDashboardPrivate
-        ? dashboardMainPageGetters.privateDashboardItems.find((item) => item.dashboard_id === treeData.id)
-        : dashboardMainPageGetters.publicDashboardItems.find((item) => item.dashboard_id === treeData.id);
-    const _dashboardWidgets = await listDashboardWidgets(_dashboard.dashboard_id);
-    const _createdLayouts = await getSharedDashboardLayouts(_dashboard.layouts, _dashboardWidgets, storeState.costDataSource);
-    const _createdDashboardParams: DashboardCreateParams = {
-        name: getClonedName(dashboardMainPageGetters.existingDashboardNameList, _dashboard.name),
-        layouts: _createdLayouts,
-        options: _dashboard.options || {},
-        labels: _dashboard.labels || [],
-        tags: { created_by: store.state.user.userId },
-        folder_id: folderId,
-    };
-    if (storeState.isAdminMode) {
-        (_createdDashboardParams as PublicDashboardCreateParameters).resource_group = RESOURCE_GROUP.DOMAIN;
-    } else if (!isPrivate) {
-        (_createdDashboardParams as PublicDashboardCreateParameters).resource_group = _dashboard?.resource_group || RESOURCE_GROUP.WORKSPACE;
-    }
-    const createdDashboard = await dashboardStore.createDashboard(_dashboardType, _createdDashboardParams);
-    dashboardMainPageStore.setNewIdList([
-        ...dashboardMainPageState.newIdList,
-        createdDashboard.dashboard_id as string,
-    ]);
 };
 
 /* Event */
@@ -148,13 +75,14 @@ const handleCloneConfirm = async () => {
     await Promise.allSettled(state.selectedTreeData.map(async (item) => {
         const _isPrivate = state.privateMap[item.data.id];
         if (item.data.type === 'FOLDER') {
-            const createdFolderId = await createFolder(item.data.name, _isPrivate);
+            const createdFolderId = await dashboardStore.createFolder(item.data.name, _isPrivate);
             if (!createdFolderId) return;
+            dashboardMainPageStore.setNewIdList([...dashboardMainPageState.newIdList, createdFolderId]);
             item.children.forEach((child) => {
-                _createDashboardPromises.push(createDashboard(child.data, _isPrivate, createdFolderId));
+                _createDashboardPromises.push(cloneDashboard(child.data.id, _isPrivate, createdFolderId));
             });
         } else {
-            _createDashboardPromises.push(createDashboard(item.data, _isPrivate));
+            _createDashboardPromises.push(cloneDashboard(item.data.id, _isPrivate));
         }
     }));
     const _results = await Promise.allSettled(_createDashboardPromises);
